+++ client/src/views/AzureAd.vue
@@ -17,8 +17,8 @@
             <div class="column">
               <BulmaCheckbox checktype="checkbox" v-model="azuread.enable" label="Enable MS Entra ID" />
               <div class="mt-2">
-                <BulmaInput :disabled="!azuread.enable" icon="user-tag" v-model="azuread.client_id" label="Client Id" placeholder="" :required="true" :hasError="$v.azuread.client_id.$invalid" :errors="[]" />
-                <BulmaInput :disabled="!azuread.enable" icon="user-secret" v-model="azuread.secret_id" type="password" label="Secret Id" placeholder="" :required="true" :hasError="$v.azuread.secret_id.$invalid" :errors="[]" />
+                <BulmaInput :disabled="!azuread.enable" icon="user-tag" v-model="azuread.client_id" label="Client Id" placeholder="" :required="true" :hasError="v$.azuread.client_id.$invalid" :errors="[]" />
+                <BulmaInput :disabled="!azuread.enable" icon="user-secret" v-model="azuread.secret_id" type="password" label="Secret Id" placeholder="" :required="true" :hasError="v$.azuread.secret_id.$invalid" :errors="[]" />
                 <BulmaInput :disabled="!azuread.enable" icon="filter" v-model="azuread.groupfilter" label="Groupname Regex" placeholder="A regular expression to match groups" :required="false" :errors="[]" />
                 <div class="notification is-info-light content">
                   <strong>Required API Permissions</strong><br>
@@ -44,17 +44,15 @@
   </section>
 </template>
 <script>
-  import Vue from 'vue'
+
   import axios from 'axios'
-  import Vuelidate from 'vuelidate'
   import BulmaButton from '../components/BulmaButton.vue'
   import BulmaInput from '../components/BulmaInput.vue'
   import BulmaCheckbox from '../components/BulmaCheckRadio.vue'
   import TokenStorage from '../lib/TokenStorage'
   import BulmaSettingsMenu from '../components/BulmaSettingsMenu.vue'
-  import { requiredIf } from 'vuelidate/lib/validators'
-
-  Vue.use(Vuelidate)
+  import { useVuelidate } from '@vuelidate/core'
+  import { requiredIf } from '@vuelidate/validators'
 
   export default{
     name: "AfAzureAd",
@@ -63,6 +61,9 @@
       isAdmin:{type:Boolean}
     },
     components:{BulmaButton,BulmaInput,BulmaCheckbox,BulmaSettingsMenu},
+    setup(){
+      return { v$: useVuelidate() }
+    },
     data(){
       return  {
           azuread:{
@@ -101,7 +102,7 @@
           };
       },updateAzureAd(){
         var ref= this;
-        if (!this.$v.azuread.$invalid) {
+        if (!this.v$.azuread.$invalid) {
           axios.put(`${process.env.BASE_URL}api/v1/azuread/`,this.azuread,TokenStorage.getAuthentication())
             .then((result)=>{
               if(result.data.status=="error"){